Split diagnose_insufficient into its own module
I reckon scheduler.py was getting a bit unmanageably large, and I want
to next split out more helper functions for the diagnose_insufficient
functionality.

To make this work I had to move a bunch of functions from
test/test_scheduler.py to test/utils.py.
